ci: fix workflows and harden KMP sync
Release bumps real module VERSION_NAME; per-module tags.
Dokka uses doc/ + HTML deploy. CI runs tests, Wasm, iOS.
Sync: wasm-safe format/exceptions; recurrence Int→Long.

+++ kalendar-sync/src/commonMain/kotlin/com/himanshoe/kalendar/sync/KalendarRecurrenceExpander.kt
@@ -64,7 +64,7 @@ fun KalendarSyncEvent.expandOccurrences(
     val occurrences = mutableListOf<KalendarSyncEvent>()
     var occurrenceCount = 0
     var current = date
-    val startEpoch = date.toEpochDays()
+    val startEpoch = date.toEpochDays().toLong()
 
     // Guard against pathological rules that never advance
     var safety = 0
